+++ game/g_shield.cpp
@@ -16,7 +16,7 @@ const object_type shield::_type(subsystem::_type);
 physics::material shield::_material(0.0f, 0.0f);
 
 //------------------------------------------------------------------------------
-shield::shield(physics::shape const* base, game::ship* owner, int compartment)
+shield::shield(physics::shape const* base, game::ship* owner, uint16_t compartment)
     : subsystem(owner, compartment, {subsystem_type::shields, 2})
     , _base(base)
     , _strength(2)
