+++ game/g_subsystem.h
@@ -72,7 +72,7 @@ public:
     static const object_type _type;
 
 public:
-    subsystem(game::ship* owner, int compartment, subsystem_info info);
+    subsystem(game::ship* owner, uint16_t compartment, subsystem_info info);
 
     virtual object_type const& type() const override { return _type; }
     virtual void think() override;
@@ -83,7 +83,7 @@ public:
     float damage() const { return _damage; }
     void repair(float damage_per_second);
 
-    int compartment() const { return _compartment; }
+    uint16_t compartment() const { return _compartment; }
 
     int current_power() const;
     int desired_power() const { return _desired_power; }
@@ -94,7 +94,7 @@ public:
 protected:
     subsystem_info _subsystem_info;
 
-    int _compartment;
+    uint16_t _compartment;
 
     float _damage;
 
@@ -127,7 +127,7 @@ public:
     static const object_type _type;
 
 public:
-    engines(game::ship* owner, int compartment, engines_info info);
+    engines(game::ship* owner, uint16_t compartment, engines_info info);
 
     virtual object_type const& type() const override { return _type; }
     virtual void think() override;
