+++ game/g_weapon.cpp
@@ -108,7 +108,7 @@ std::vector<weapon_info> weapon::_types = {
 };
 
 //------------------------------------------------------------------------------
-weapon::weapon(game::ship* owner, int compartment, weapon_info const& info, vec2 position)
+weapon::weapon(game::ship* owner, uint16_t compartment, weapon_info const& info, vec2 position)
     : subsystem(owner, compartment, {subsystem_type::weapons, 2})
     , _info(info)
     , _last_attack_time(time_value::zero)
